+++ .gitignore
@@ -90,3 +90,5 @@ dist
 # Temporary folders
 tmp
 temp
+
+*.foxe
