+++ package.json
@@ -9,13 +9,13 @@
   "main": "./dist/extension.js",
   "keywords": [],
   "scripts": {
-    "build": "foxglove-extension build",
-    "foxglove:prepublish": "foxglove-extension build --mode production",
+    "build": "lichtblick-extension build",
+    "lichtblick:prepublish": "lichtblick-extension build --mode production",
     "lint:ci": "eslint --report-unused-disable-directives .",
     "lint": "eslint --report-unused-disable-directives --fix .",
-    "local-install": "foxglove-extension install",
-    "package": "foxglove-extension package",
-    "pretest": "foxglove-extension pretest"
+    "local-install": "lichtblick-extension install",
+    "package": "lichtblick-extension package",
+    "pretest": "lichtblick-extension pretest"
   },
   "devDependencies": {
     "@foxglove/eslint-plugin": "1.0.1",
@@ -24,7 +24,7 @@
     "@types/react-dom": "18.2.17",
     "@typescript-eslint/eslint-plugin": "6.13.1",
     "@typescript-eslint/parser": "6.13.1",
-    "create-foxglove-extension": "0.8.6",
+    "create-lichtblick-extension": "1.0.0",
     "eslint": "8.54.0",
     "eslint-config-prettier": "8.10.0",
     "eslint-plugin-es": "4.1.0",
@@ -40,6 +40,6 @@
     "typescript": "5.3.2"
   },
   "dependencies": {
-    "@foxglove/schemas": "^1.6.2"
+    "@foxglove/schemas": "^1.8.0"
   }
 }
