+++ src/msgs/DetectedObjects.ts
