+++ src/msgs/Header.ts
@@ -1,0 +1,6 @@
+import { Time } from "@foxglove/schemas";
+
+export type Header = {
+  stamp: Time;
+  frame_id: string;
+};
